+++ src/targetValidationUtils.js
